[Model] Address LTX-2.5 review feedback
